requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pypdf
docx2txt